Fix linkage to vigranumpy on Mac.
For some reason, the linkage to vigranumpy gets pointed at the wrong
path. To fix this, we go ahead and run `install_name_tool` on Mac to
change the path back to the absolute path. As `conda-build` can go back
through and fix this for us, we go ahead and let it do this to get back
to a correct, relative RPATH. This is a different approach to a similar
issue we experienced on Linux some time ago.