Support Python 3.4
As many other libraries that we maintain support Python 3.4, it seems
like a good idea to verify that this also can be built and installed on
Python 3.4. The main interest in this version of Python is that it was
supported by conda-forge for a while and some packages of rank_filter
were built on it. This will help verify that this does in fact work
correctly on Python 3.4 and keep that target in mind for the future.