Optionally require VIGRA for extra tests
In order to run some of the C++ tests, we require VIGRA as a build
dependency. However there are not builds of VIGRA for all combinations
of Python and NumPy that we support. So we may not be able to build and
test all of those cases. To address this in such a way where we can run
the tests as needed, we add an environment variable to allow us to
determine whether VIGRA should be included in the build (by default it
is disabled). This way we can selectively run these tests.