Add VIGRA-based C++ tests to the build matrix
Selectively enable VIGRA-based C++ tests for one case in the build
matrix. As these tests do not depend on Python or NumPy in any form,
they only need to be run in one matrix case in order to be sufficiently
tested. So VIGRA is only require in one of the matrix cases and left out
from all other cases.