Drop old check for undefined environment variables
It use to be the case that `conda-build` would define undefined
environment variables with the `<UNDEFINED>` string. However this was
fixed in `conda-build` 1.19.1. Since conda-build 1 is no longer
supported, and `conda-build` 2 and 3 do not have this problem (where
`conda-build` 2 is basically ubiquitous), it make sense to drop this
outdated check.