Use Boost's round
It appears the code in the base header file is relying on a `round`
implementation to be pulled in by something else. However this may be an
implementation that requires C++11 support. To correct this, make use of
Boost's `round`. This should be guaranteed to work on incredibly old
compilers and avoid accidentally adding a C++11 hard requirement.