Standardize types between both iterators
Make sure that both iterators take the same value type. This constraint
is added to ensure that there is no conversion being done by running the
filter. Once this is verified define this value type in an iterator
independent way. Also find an appropriate difference type between the
two iterators that will work in both cases.