Preallocate deque used for the filter window
This is needed on some platforms to ensure that the needed space is
already allocated before adding values (e.g. Windows). That said, this
should be a bit faster than adding values to the filter window at the
time and potentially going through several reallocations in the process.
By using this approach, we guarantee that only one allocation occurs
when creating and initially filling the deque. Ideally there should be
no reallocations in the filter code as we keep this size as an upper
bound (though we do go under by 1 periodically).