Eliminate `lambda` wrappers
Instead of using `lambda` wrappers`, simply call the fused type function
directly with the intended specialization. Should cutdown the overhead
of dispatch and make it easier for the compiler to inline it. This
remains pretty compact thanks to the conversion of the function to an
in-place computation and some previous refactoring.