Enforce `image` and `out` are `ndarray`s
Make sure that `image` and `out` are typed as `ndarray`s in the function
signature. Smooth over a few spots that don't work out-of-the-box by
coercing them to `object`s explicitly. While there is still that can be
done to the lower the code in `lineRankOrderFilter` close to C, adding
this typing makes it easier to pursue internally without breaking API.