Implement an `ndindex`-like Cython function
Provides a Cython function that acts similarly to `ndindex`. However it
is written in C-style. So it does not yield, does not use classes, etc.
It also doesn't keep state. However it can be used like a generator
except in a `while`-loop where state external to the function is update
on each run.