Use float pointers in Cython/C++ binding function
Change the Cython/C++ binding function,
`lineRankOrderFilter1D_floating_inplace`, to accept floating pointers
instead of memoryviews. This cuts down the code generated in the
`while`-loop to convert the pointer to a memoryview and in the
Cython/C++ binding function. Should make it easier for compilers to
optimize further as the code is basically straight C at this point. Also
should help speed up the `while`-loop a bit more.