Drop use of `memoryview`s in our `while`-loop func
As we are only using straight pointers within the `while`-loop, there is
little value to creating `memoryview`s. Especially since we are going
through an extra step to cast our pointers as `memoryview`s. Not to
mention creating them requires the GIL. So instead just cutout
`memoryview`s and use pointers directly. Should improve the performance
a bit as well.