Rewrite looped filter function using a pointer
Instead of making use of things like a custom Cython version of
`ndindex` and NumPy's `PyArray_GetPtr`, rewrite the looped filter
function to simply take a pointer into the data and then step through
the data in increments of the last axis' length. Determine the number
1-D slices like this that the array contains and ensure each one is
visited. This significantly simplifies the code and improves
performance.