Call C++ function directly from looped filter func
As there really is not much more to just calling the C++ rank filter
function directly from our looped filter function, go ahead and update
it accordingly. Avoids some function call overhead and avoids the need
to have the compiler do the inlining for us as we have already done it
effectively.