Rewrite window initialization with source iterator
Uses the source iterator to do a single pass of the data in source.
Inserts values in order as they would appear starting with the reflected
values using a deque to buffer and orientate the sequence and then
inserts them to the multiset with iterators of these values going into
deque. This ensures the in order sequence property of the data is
preserved, which is needed for the rolling algorithm to handle removal
correctly.