Eliminate need for finding the container length
Rewrite `while`-loop to make use of iterators to determine when
reflection needs to occur and when the filter has reached the end. Keep
track of the mapping between values within the window and their
reflected counterparts once the filter crosses outside of the source
container. With this change, there is no need to know the length of the
source or the destination (except for assertion purposes). So drop those
steps.