Use `std::advance` to get `rank_point`
While iteratively incrementing the iterator, `rank_point`, is a valid
way of getting it to the intended position, `std::advance` can do the
same thing and is a more compact way of writing this. Plus if the
iterator is a `RandomAccessIterator`, `std::advance` can jump to the
intend position in constant time. As we are typically using this with
pointers into a contiguous array, we benefit from this special case
optimization.