Store only reflected data in deque
This halves the initial window data's deque and only stores the data
reflected in it. Does a forward pass of the initial window's deque to
add the reflected data into the multiset with iterators stored in order
in the rolling window's deque. Then adds the data to the multiset in
reverse order to handle the unreflected portion. Pops values from the
initial window's deque on the way.