Consolidate window end reflection cases
The last pass of the filter wasn't strictly necessary as its result
wasn't actually being stored in the destination. To fix this change the
`while`-loop check to verify that the source container still has more
values or the window still has values to reflect to complete the pass
over the source data. As one of these must always be true, there is no
need to handle the case where neither of these are true. After the final
pass, make sure to store the last result in the destination before
exiting.