Track whether the source container has data still
Keep a record of whether the source container still has data. That way
this can be checked only once per `while`-loop iteration and only
updated if the source iterator was updated. As soon as it is determined
the source container is out of data, note this and do not perform
further checks as it remains exhausted.