Use `size_t` for all indices
Previously the code used difference types for the respective iterators.
This made sense when the code still took differences between iterators
and used these differences to implement the moving window. However the
code has since been rewritten to support forward iterators, and all
remaining indices are connected to the window length, which is `size_t`.
Hence all indices are switched to `size_t`.