Drop Windows 32-bit
With as many jobs this takes, the fact that the builds each take a bit,
and only one build runs at a time, we need to be a bit more economical
with which builds we choose to run. In particular, it doesn't make sense
to continue supporting Windows 32-bit. We don't use 32-bit.  Also the
user base of 32-bit has been dwindling generally (as shown by a
conda-forge survey). Removing 32-bit would halve the number of builds.
Given all of this, dropping 32-bit seems like the right move and so we
do it here.